Move __init__ parse_func functions into it's own file.
